refactor(BEB-23): move configuracoes do swagger para o diretorio shared
